drm: samsung: fixed wrong shadow mode configuration of dsim
The reason why cprs_en register could not be set to 1 in previous
dsim initial code is that cprs_en is set to 1 after shadow mode
is enabled but shadow mode is disabled without shadow update immediately.
So, cprs_en configuration could not be applied to hw.

This removes to disable shadow mode of dsim. The dsim hw will
always work in shadow mode.

Bug: 179475534
Signed-off-by: Jiun Yu <jiun.yu@samsung.com>
Change-Id: I61ceb58ec36f7a2089f47236dab47c1a2d4d3743