s6e3hc3: bundle brightness update with HBM update
HBM and brightness updates are controlled in different sysfs nodes. It
is possible that HBM and brightness update does not happen in the same
frame and cause panel flicker if there is any event in between such as
self refresh rate update between HBM off and brightness update.  Bundle
brightness update with HBM on/off to secure HBM and brightness update in
the same frame.

Bug: 197479033
Signed-off-by: William Tai <williamtai@google.com>
Change-Id: I5e94a7284e24240518c018faa059bc42e7a3823e