GKI: drm: samsung: introduce EXYNOS_DSI_MSG_QUEUE flag
Commit 1026fb2f81df1 ("ANDROID: drivers: gpu: drm: add support to batch
commands)" was reverted in ACK breaking the ability to batch commands
sucessfully. Since MIPI_DSI_MSG_LASTCOMMAND flag was no longer set on
mipi_dsi_device_transfer(), the commands would be kept queued until the
number of pending packets exceeded its threshold. This resulted in a
visible lag on the display (specially when UART was enabled).

This patch introduces the EXYNOS_DSI_MSG_QUEUE flag as an inverse logic
replacement of the reverted MIPI_DSI_MSG_LASTCOMMAND flag. Commands not
marked with EXYNOS_DSI_MSG_QUEUE flag should now be considered the last
command of the batch and be sent immediately.

Bug: 201090883
Test: Batching works. Display lag fixed even with UART enabled.
Suggested-by: Adrian Salido <salidoa@google.com>
Signed-off-by: Carlos Llamas <cmllamas@google.com>
Change-Id: I2662d96e66a21c38567e46ea36bab54f03355048