panel: nt37290: avoid noise lines in 60Hz
There are underruns in 60Hz because the time for DPU start write
data to GRAM is not ahead enough before DDIC start read data from
GRAM. This change shifts TE about 8.2ms so that DPU can start to
write data early to avoid the problem. While switching to 120Hz, the
shift cannot be used, we have to send the command to restore the
TE timing.

Bug: 203626904
Test: no noise lines in 60Hz and 120Hz
      switch between 60Hz and 120Hz without problems
      suspend/resume in 60Hz or 120Hz
Change-Id: Ifbf8382284cfb659f341644a0b4977c32ef3106d
Signed-off-by: Chris Lu <luchris@google.com>