panel: nt37290: modify LHBM init command
1. Modify panel revision condition
The commands are needed for EVT1 and after, and they are OTPed
since EVT1.1. That is, we only need to send the commands for EVT1.

2. Modify DBV for proto1.1
The last commands are only needed for proto1.1, but these will
cause brightness brighter for a while in the boot animation. Modify
the DBV to avoid the problem.

Bug: 228650741
Test: boot to home without peak brightness
      enable LHBM without problems
Change-Id: Ia01836e5de047cf556826c2537d139a9e5bba236
Signed-off-by: Chris Lu <luchris@google.com>