drm: samsung: add support of delay DSC reg init
Add a variable to support delay DSC reg init. It waits for the
next TE rising and extra delay to make sure the frame start
happens after both DPU and panel PPS are set.

The delay time could be defined in each panel file. If the value
is 0 or it's not defined, the DSC reg init will be called during
the original sequence.

Bug: 228383688
Test: capture a trace to check the sequence
Change-Id: I40ee247cfe0d2aabedda3cac49953f44616a356d
Signed-off-by: Chris Lu <luchris@google.com>