panel: nt37290: allow OSC2 access while in AOD and OFF
We may change the OSC2 clock while in AOD and OFF states, but
the commands cannot be sent to DDIC except in normal state. Add
the change to allow the access but just save the value. It will
be applied when the state is changed to normal.

If some panel cannot allow the access, the protection can be
added in the implementation in each panel file.

Bug: 212601547
Test: read/write the node while in AOD and OFF states
Change-Id: I49d8f54eca20b09087868dbd1d727b9687ddf894
Signed-off-by: Chris Lu <luchris@google.com>