panel: nt37290: fix the calculation of DVT brightness
For DVT brightness, we obtain the nits via the interpolation
since the original curve is linear. Then we get the new DBV
from nonlinear equations. However, the original curve is separated
to NBM and HBM with different slopes. Thus we have to separate
the calculation to NBM and HBM with different parameters, i.e.
min and max values to get correct results from interpolation.

Bug: 231899714
Test: check brightness logs, HW measurement
Change-Id: Ic66baec112db4be52092043f3cb902f68be35580
Signed-off-by: Chris Lu <luchris@google.com>