drm: samsung: split out common API code from the cal_* drivers
The drivers exynos_drm_dpp.c and exynos_drm_dqe.c call into the cal_*
drivers via a common interface that is shared between the 9845, 9855,
and 9865 drivers. Currently 9845 and 9855 are compiled together and the
9855 driver calls into the 9845 driver. However, the 9865 driver doesn't
depend on the 9845 or 9865 drivers and redefines some of the same APIs.
So this patch moves the common entry functions that the exynos drm and
dqe drivers call to the cal_common/* drivers. It also updates the Kbuild
file so that we can compile either the 9845 and 9855 drivers or the 9865
drivers so that there aren't build errors related to re-defining
symbols.

Signed-off-by: Will McVicker <willmcvicker@google.com>
Change-Id: I4dff41ad8f0c54be288863ec436c97ace0cb5b42