drm: samsung: protect to init and deinit DSIM/DPHY hardware.
Multiple DSIM and DPHY can be initialized or deinitialized
at the same time. So, it causes problem.

For example, if DSIM0 and DSIM1 are initialized at the same time,
PLL initialization can be failed because of shared SFRs
and DSIM/DPHY hardware initialization sequence.

So, DSIM global mutex lock protects multiple DSIM/DPHY initialization
or deinitialization.

Change-Id: I5ada2e686e5f7dd340afd311650c8982f0db6362
Signed-off-by: YongWook Shin <yongwook.shin@samsung.com>
Signed-off-by: Jiun Yu <jiun.yu@samsung.com>