drm: samsung: dsim: cleanup packet-go even after failure
When there's a timeout sending a series of commands using packet-go
feature, sending of commands could fail due to unexpected issues.
In such cases we still need to ensure that the state of the feature is
cleared even if it timed out, otherwise we may land in unexpected sw
state.

Bug: 269702366
Test: turn screen on/off, check dynamic debug logs
Change-Id: I65648964372daa1519fbe68d796425f4bb5cf56b
Signed-off-by: Adrian Salido <salidoa@google.com>