drmdp: Expose hotplug error code through sysfs
DisplayPort kernel driver notifies hardware composer (HWC) running in
user-space by sending a hotplug uevent. In case of failure we also need
to tell HWC the error code, but the hotplug uevent can't be easily
modified (it's defined in upstream DRM KMS code), so we are adding a
sysfs node. HWC will read the numerical error code from that file and
will write "0" to it, to reset the error code after it is consumed.

Test: manual
Bug: 283461313
Signed-off-by: Sergey Volk <servolk@google.com>
Change-Id: Iee9d783e6bd451a6829775f77d0cb8a830e89a8b